🐛(api) forbid `/static/` POST usage for an existing PDC
If a user submits the same payload twice using the /statique/ (POST)
endpoint, the API returns a 201 CREATED. And if the payload is slightly
modified (with the same id_pdc_itinerance), we cannot guaranty what
would be created or updated. We now return a 409 CONFLICT response if
the PDC already exist.