🗃️(api) make location address/coordinates unique together
Instead of having a unique `Localisation.adresse_station` field, let's
be less pedantic and allow address to be unique while combined with
specific coordinates. This allows to define a sparse address for a
station with precise coordinates. Two stations located in the same area
with a partial address will have two separated Localisation entries when
their coordinates differ.