🗃️(api) migrate database enum names to values (1)
By default, python-derived enum fields are stored using the enum name
instead of its value in database. This forces us to map the
corresponding value when exporting data.

In this first step of the migration, we create the new enum types (using
values) and then alter columns to the legacy enums to a varchar.