🧑‍💻(api) use compose watch only for rebuild actions
The `sync` action is not bi-directional: it only occurs from the host to
the container, hence using a formatter in the container (e.g. black)
will not update the host original file. We better switch back to a
classical volume and use watch feature to trigger a built when updating
dependencies.