🦺(api) restrict submitted session and status max age
For legal reasons, we need to forbid sessions older than 15 days. Fow
now, we set this limit to a year, but it will be reconfigured to be set
to its expected limit in January 2026.

We applied the same pattern to statuses to avoid receiving useless
outdated statuses: statuses older than one-day are no longer accepted.